Backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Amirdhesh <amirdhesh2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
sqlmodel = "^0.0.14"
fastapi = "^0.109.0"
alembic = "^1.13.1"
black = "^23.12.1"
pytest = "^7.4.4"
pydantic-settings = "^2.1.0"
pymysql = "^1.1.0"
uvicorn = "^0.27.0.post1"
pydantic = {extras = ["email"], version = "^2.6.0"}
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
